+++ hammer-api/src/test/java/cn/featherfly/hammer/dml/builder/sql/update/UserUpdate.java
@@ -6,6 +6,7 @@ import java.util.function.Consumer;
 import cn.featherfly.common.lang.function.SerializableFunction;
 import cn.featherfly.common.lang.function.SerializableSupplier;
 import cn.featherfly.hammer.dml.builder.sql.query.UserQueryConditionGroupExpression;
+import cn.featherfly.hammer.expression.condition.ConditionGroupConfig;
 import cn.featherfly.hammer.expression.execute.SetExecutableUpdateExpression;
 
 /**
@@ -157,15 +158,6 @@ public class UserUpdate implements UserPropertiesUpdate,
      * {@inheritDoc}
      */
     @Override
-    public UserQueryConditionGroupExpression where(Consumer<UserQueryConditionGroupExpression> consumer) {
-        // YUFEI_TODO Auto-generated method stub
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
     public UserUpdate set(Consumer<UserUpdate> consumer) {
         // YUFEI_TODO Auto-generated method stub
         return null;
@@ -176,6 +168,15 @@ public class UserUpdate implements UserPropertiesUpdate,
      */
     @Override
     public UserUpdate increase(Consumer<UserUpdate> consumer) {
+        // YUFEI_TODO Auto-generated method stub
+        return null;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public UserQueryConditionGroupExpression where(Consumer<ConditionGroupConfig> consumer) {
         // YUFEI_TODO Auto-generated method stub
         return null;
     }